TestNinja/Mocking/VideoService.cs
using System;
using System.Collections.Generic;
using System.Data.Entity;
using System.IO;
using System.Linq;
using Newtonsoft.Json;

namespace TestNinja.Mocking
{
    public class VideoService
    {
        private readonly IFileReader fileReader;
        private readonly IVideoRepository videoRepository;
        

        public VideoService(IFileReader fileReader, IVideoRepository videoRepository)
        {
            this.fileReader = fileReader;
            this.videoRepository = videoRepository;
        }

        public string ReadVideoTitle()
        {
            var str = fileReader.Read("video.txt");

            var video = JsonConvert.DeserializeObject<Video>(str);
            if (video == null)
                return "Error parsing the video.";
            return video.Title;
        }

        public string GetUnprocessedVideosAsCsv()
        {
            var videos = videoRepository.GetUnprocessedVideos();

            var videoIds = videos.Select(v => v.Id).ToList();

            return string.Join(",", videoIds);
        }
    }

    public class Video
    {
        public int Id { get; set; }
        public string Title { get; set; }
        public bool IsProcessed { get; set; }
    }

    public class VideoContext : DbContext
    {
        public DbSet<Video> Videos { get; set; }
    }
}